@@ -5,7 +5,7 @@ homepage: https://github.com/woodemi/quick_blue
 repository: https://github.com/woodemi/quick_blue/tree/master/quick_blue
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0 <2.0.0"
 
 dependencies:
@@ -17,7 +17,7 @@ dependencies:
     path: ../quick_blue_macos
   quick_blue_windows:
     path: ../quick_blue_windows
-  convert: ^2.1.1
+  convert: ^3.0.1
 
 dev_dependencies:
   flutter_test:
