@@ -6,7 +6,7 @@ description: Demonstrates how to use the quick_blue plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,8 +22,8 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  convert: ^2.1.1
+  cupertino_icons: ^1.0.3
+  convert: ^3.0.1
 
 dev_dependencies:
   flutter_test:
